third_party/WebKit/LayoutTests/typedcssom/perspectiveTransformComponent.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(function() {
  var calcLength = new CalcLength({px: 10, percent: 3.2});
  assert_throws(null, function() { new Perspective(calcLength) });
}, "Constructor should throw an error for CalcLengths with a percentage type");

test(function() {
  var simpleLength = new SimpleLength(10, 'percent');
  assert_throws(null, function() { new Perspective(simpleLength) });
}, "Constructor should throw an error for SimpleLengths with a percentage type");

test(function() {
  var simpleLength = new SimpleLength(10, 'px');
  var calcLength = new CalcLength({px: 10, em: 3.2});
  var perspectiveTransformSimple = new Perspective(simpleLength);
  var perspectiveTransformCalc = new Perspective(calcLength);

  assert_equals(perspectiveTransformSimple.cssString, 'perspective(10px)');
  assert_equals(perspectiveTransformCalc.cssString,'perspective(calc(3.2em + 10px))');
}, "cssString should return a string of form perspective(<LengthValue.cssString()>)");

</script>

<body>
</body>
